% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sdc.R
\name{apply_sdc}
\alias{apply_sdc}
\title{Easy helper for 'apply_sdc'}
\usage{
apply_sdc(
  data,
  level = 5,
  rounding = FALSE,
  round_val = 5,
  mask = as.numeric(NA)
)
}
\arguments{
\item{data}{the data to apply SDC to}

\item{level}{the value below which SDC will be applied}

\item{rounding}{whether to apply rounding}

\item{round_val}{the value to round to}

\item{mask}{the value to mask values with}
}
\description{
Apply SDC to a data set
}
