% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_chart_hc.R
\name{basic_chart_hc}
\alias{basic_chart_hc}
\title{Easy helper for basic chart in highcharter}
\usage{
basic_chart_hc(
  data,
  x,
  y,
  type = "line",
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  seriesName = "Series 1",
  color = "#005EB8",
  dlOn = TRUE,
  currency = FALSE,
  alt_text = NULL
)
}
\arguments{
\item{data}{the data object to pass to the chart}

\item{x}{the column in the data to show on the x axis}

\item{y}{the column in the data to show on the y axis}

\item{type}{the highcharter chart type to use (defaults to line)}

\item{xLab}{the label to use on the x axis (defaults to NULL)}

\item{yLab}{the label to use on the y axis (defaults to NULL)}

\item{title}{the chart title (defaults to NULL)}

\item{seriesName}{the Series name (defaults to Series 1)}

\item{dlOn}{TRUE/FALSE whether to include data labels on the chart (defaults to TRUE)}

\item{currency}{TRUE/FALSE whether chart y values are currency (GBP) (defaults to TRUE)}

\item{marker}{TRUE/FALSE enable markers in chart (defaults to TRUE)}
}
\description{
Create basic chart in highcharter
}
\examples{
group_chart_hc()
 function to build a basic chart with one series using highcharter
 
 data labels are formatted by custom JS functions that allow for currency

}
