% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{create_metadata}
\alias{create_metadata}
\title{Easy helper for 'create_metadata'}
\usage{
create_metadata(workbook, fields, descriptions)
}
\arguments{
\item{workbook}{the name of the workbook object created using the create_wb function}

\item{fields}{a list of field names for each item of metadata}

\item{description}{a list of descriptions for each item of metadata}
}
\description{
Create metadata tab for named workbook
}
\examples{
meta_fields <- c("BNF Chapter Code",
                 "BNF Chapter Name",
                 "BNF Section Code",
                 "BNF Section Name",
                 "Cost"
)
meta_descs <- c("The unique code used to refer to the British National Formulary (BNF) chapter.",
                "The name given to a British National Formulary (BNF) chapter. This is the broadest grouping of the BNF therapeutical classification system.",
                "The unique code used to refer to the British National Formulary (BNF) section.",
                "The name given to a British National Formulary (BNF) section. This is the next broadest grouping of the BNF therapeutical classification system after chapter.",
                "There are many costs incurred when a dispensing contractor fulfils a prescription. In primary care the costs reported in this publication represent the basic price of the item. This is sometimes called the 'Net Ingredient Cost' (NIC). This also known as reimbursement of costs to dispensing contractors. In secondary care they are the actual costs paid (including applicable VAT) for drugs, dressing, appliances, and medical devices which have been issued and used in NHS hospitals in England."
)
create_metadata(test,
                meta_fields,
                meta_descs
)
}
