% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wb.R
\name{create_wb}
\alias{create_wb}
\title{Easy helper for 'create_wb'}
\usage{
create_wb(sheets)
}
\arguments{
\item{sheets}{a list of sheet names require}
}
\description{
Create base workbook 'wb' and add sheets required
}
\examples{
sheetNames <- c("test1","test2","test3")
create_wb(sheetNames)
}
