% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Easy helper for 'format_data'}
\usage{
format_data(workbook, sheetname, column, alignment, number_format)
}
\arguments{
\item{workbook}{the name of the workbook object created using the create_wb function}

\item{sheetname}{the name of the sheet to apply style to}

\item{column}{the excel column letter(s) you wish to apply the style to eg. "F" or c("A", "B")}

\item{alignment}{Horizontal alignment of cell contents
\itemize{
  \item \strong{left} Left horizontal align cell contents
  \item \strong{right} Right horizontal align cell contents
  \item \strong{center} Center horizontal align cell contents
  \item \strong{justify} Justify horizontal align cell contents
}}

\item{number_format}{the Excel number format code you wish to apply}

\item{filepath}{the file path to save the workbook to}
}
\description{
Format data on selected sheet of base 'wb' created using create_wb function
}
\examples{
format_data(myworkbook,
"test1",
"A",
"right",
"#,###"
)
format_data(myworkbook,
"test1",
c("A", "B", "C"),
"right",
"#,###.00"
)
format_data(myworkbook,
"test1",
C("D", "AA", "X"),
"left",
""
)
}
