% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_chart_hc.R
\name{group_chart_hc}
\alias{group_chart_hc}
\title{Easy helper for group chart in highcharter}
\usage{
group_chart_hc(
  data,
  x,
  y,
  type = "line",
  group,
  xLab = NULL,
  yLab = NULL,
  title = NULL,
  dlOn = TRUE,
  currency = FALSE,
  marker = TRUE
)
}
\arguments{
\item{data}{the data object to pass to the chart}

\item{x}{the column in the data to show on the x axis}

\item{y}{the column in the data to show on the y axis}

\item{type}{the highcharter chart type to use (defaults to line)}

\item{group}{the column in the data which includes the groups to use in the chart}

\item{xLab}{the label to use on the x axis (defaults to NULL)}

\item{yLab}{the label to use on the y axis (defaults to NULL)}

\item{title}{the chart title (defaults to NULL)}

\item{dlOn}{TRUE/FALSE whether to include data labels on the chart (defaults to TRUE)}

\item{currency}{TRUE/FALSE whether chart y values are currency (GBP) (defaults to TRUE)}

\item{marker}{TRUE/FALSE enable markers in chart (defaults to TRUE)}
}
\description{
Create group chart in highcharter
}
\examples{
group_chart_hc()
}
