% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{not_in}
\alias{not_in}
\alias{\%!in\%}
\title{Easy helper for 'not in'}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\description{
Filter vectors that are not in a given vector using syntax similar to the base
R `%in%` operator.
}
\examples{
x <- c(1, 2, 3, 4, 5)
x[x \%!in\% c(1, 3, 5)]
}
