% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Fast CSV writer}
\usage{
save_data(data, dir = ".", filename, ...)
}
\arguments{
\item{data}{A data.frame object to be written to CSV}

\item{dir}{The directory to save the CSV to. defaults to `.` - current
working/project directory. Creates `data` sub-directory if it doesn't already
exist}

\item{filename}{Name for the file. This is concatenated with a time stamp
for audit purposes.}

\item{...}{Optional arguments to be passed to `data.table::fwrite`.}
}
\value{
`NULL`
}
\description{
Write data to storage using `data.table::fwrite`. This function
also adds a time stamp to the file name to allow auditing of data files.
}
\examples{
save_data(mtcars, filename = "mumh_quarterly")

save_data(
mtcars,
dir = "Y:/Official Stats/MUMH",
filename = "mumh_quarterly",
quote = TRUE
)
}
