% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmd_data.R
\name{scmd_data}
\alias{scmd_data}
\title{Manipulate SCMD data}
\usage{
scmd_data(file)
}
\arguments{
\item{file}{}
}
\value{
a data frame
}
\description{
Manipulate SCMD data
}
\examples{
scmd_data(file = "path")
}
