% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmd_icb_bnf_data.R
\name{scmd_icb_bnf_data}
\alias{scmd_icb_bnf_data}
\title{Manipulate SCMD data}
\usage{
scmd_icb_bnf_data(file)
}
\arguments{
\item{file}{}
}
\value{
a data frame
}
\description{
Manipulate SCMD data
}
\examples{
scmd_icb_bnf_data(file = "C:/ICB_BNF_2022.xlsx")
}
