% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_dwh.R
\name{table_1_dwh}
\alias{table_1_dwh}
\title{Easy helper for 'table_1_dwh'}
\usage{
table_1_dwh(con)
}
\arguments{
\item{con}{The database connection object to be used}
}
\description{
Extract DWH data to be used in table_1
}
