% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_dwh.R
\name{table_3_dwh}
\alias{table_3_dwh}
\title{Easy helper for 'table_3_dwh'}
\usage{
table_3_dwh(con)
}
\arguments{
\item{con}{The database connection object to be used}
}
\description{
Extract DWH data to be used in table_3
}
