% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_4_dwh.R
\name{table_4_dwh}
\alias{table_4_dwh}
\title{Easy helper for 'table_4_dwh'}
\usage{
table_4_dwh(con)
}
\arguments{
\item{con}{The database connection object to be used}
}
\description{
Extract DWH data to be used in table_4
}
