% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_dwh.R
\name{table_6_dwh}
\alias{table_6_dwh}
\title{Easy helper for 'table_6_dwh'}
\usage{
table_6_dwh(con)
}
\arguments{
\item{con}{The database connection object to be used}
}
\description{
Extract DWH data to be used in table_6
}
