% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_dwh.R
\name{table_7_dwh}
\alias{table_7_dwh}
\title{Easy helper for 'table_7_dwh'}
\usage{
table_7_dwh(con)
}
\arguments{
\item{con}{The database connection object to be used}
}
\description{
Extract DWH data to be used in table_7
}
