% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_8_dwh.R
\name{table_8_dwh}
\alias{table_8_dwh}
\title{Easy helper for 'table_8_dwh'}
\usage{
table_8_dwh(con)
}
\arguments{
\item{con}{The database connection object to be used}
}
\description{
Extract DWH data to be used in table_8
}
