% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fact.R
\name{create_fact}
\alias{create_fact}
\title{Easy helper for 'create_fact'}
\usage{
create_fact(
  con,
  from = 201504L,
  to = dplyr::sql("MGMT.PKG_PUBLIC_DWH_FUNCTIONS.f_get_latest_period('EPACT2')")
)
}
\arguments{
\item{con}{The database connection object to be used}

\item{from}{The first month you wish to collect data from (defaults to 201504)}

\item{to}{The last month you wish to collect data from (defaults to maximum month in DWH)}
}
\description{
Build 'HRT_FACT_DIM' in personal schema
}
